+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added `aioplus.amax`;
-* Added `aioplus.amin`.
+* Added `aioplus.amin`;
+* Added `aioplus.aminmax`.
 
 ## [0.2.0] - 2025-08-19